Creating new utils class
